requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
certifi==2024.2.2
solana==0.31.0
solders==0.19.0
base58==2.1.1 